+++ src/vehicle_system/autoaim/armor_detector/src/armor_detector/armor_detector.cpp
@@ -2,7 +2,7 @@
  * @Description: This is a ros-based project!
  * @Author: Liu Biao
  * @Date: 2022-10-13 23:26:16
- * @LastEditTime: 2023-05-05 03:05:58
+ * @LastEditTime: 2023-05-11 21:55:27
  * @FilePath: /TUP-Vision-2023-Based/src/vehicle_system/autoaim/armor_detector/src/armor_detector/armor_detector.cpp
  */
 #include "../../include/armor_detector/armor_detector.hpp"
@@ -209,7 +209,7 @@ namespace armor_detector
             }
 
             //生成装甲板旋转矩形和ROI
-            std::vector<Point2f> points_pic(armor.apex2d, armor.apex2d + 4);
+            std::vector<Point2d> points_pic(armor.apex2d, armor.apex2d + 4);
             cout << "cornor:";
             for (auto pts : points_pic)
             {
